Initial commit
So far most opcodes have been implemented. All 0xCB opcodes are done and working. memory is implemented as 1 64kb
uint8 array. bank switching for ROM and RAM is not supported yet. gpu is implemented but very buggy and basically
doesn't work yet. keyboard is supported, but has same problems as gpu. basic cpu functions are
all there except some of the halt and stop opcodes and most interrupts.
opcodes and