Added SDL_Flip!
added SDL_Flip, so changes to screen are actually visible now.
step_lcd seems to work fairly well. Initialized IME and a few timer
and div registers. Joypad should work, but roms can't get to a point
where I could test it. JR and JR_COND aren't working correctly, but
I have a list of opcodes that tetris should run and I'm currently
testing and comparing with that list to test for faulty opcodes. Changed
a few things in draw_line, but again can't tell if it's working or not
because certain opcodes aren't working correctly, so it's drawing stuff
but it's not the right stuff. Also realized that linebuffer[i] should
be number(uint8) between 0 and 3, not a color(Uint32); I should probably
remove the commented linebuffer[i] = pal[pixel]