Updated interrupts
added stop opcode(same function as halt). Fixed bug in mem.c when writing to IE. Changed a lot of interrupt stuff and they should be working now(the code was essentially the same as before, the bug in mem.c was probably the bottleneck preventing interrupts from being executed).