keypresses only trigger interrupts when the corresponding set of keys is loaded in memory
also realized I need to make a distinction between write_byte called by the rom and
and write_byte called in lcd.cpp/mem.cpp. most of the memory accesses in opcodes.h/cb_opcodes.h
are called by the rom but it seems that only some of the writes in cpu.cpp also fall into this
category. having two distinct write functions is preferable to just making memory public and
changing it directly since it creates a centralized way to debug all writes to memory