starting adding option for dynamic recompilation
added traverse() for reading in opcodes and arguments between conditional
branches. still thinking on whether I should use the native stack to implement
the foreign stack or emulate it as part of memory. added new files to Makefile.
added function to check if a block with a given start address is in the cache.
the cache is currently implemented as just an array of 32 blocks. still need
to think about the best way to implement this. added dynamic recompilation
option to argp.