started implementing MBC
when write_byte is called with address in MBC range (0x0000-0x7fff or 0xa000-0xbfff),
handle_mbc is called with the same arguments to implement any side effects of the
attempt to write. handle_mbc is a pointer to a member function of mem which is set to
handle_xxx (where xxx is mbc1/mbc2/mbc3/romonly) during load_cart(). splitting up MBC
handling this way avoids having to check the mbctype each time the program interacts
with the MBC at the expense of having to write very similar functions for each
type of MBC (about 5 well-documented types). On an unrelated note, it may be a good
idea to write an (inlined) handle_io function. Though there would only be one handle_io
function, it would improve readability in write_byte and further separate the concept
of writing to memory (while considering write permissions) from the side effects of
each write.