simplified cpu throttling and found decent parameters
only using two variables to control throttling (and two more as a step size)
Kirby now idles at about 60% cpu usage but this could still be improved.
fixed a bug in window drawing part of draw_line() related to using the wrong
tile indices. Kirby is now rendered correctly but pkblue still has issues.
I believe it's related to scrolling the window in the y direction since there
are multiple copies of the same image in some scenes in pkblue.