added RAM saving and loading
Stores binary .sav files in saves/ by default. Also attempts to load a .sav
with the same name as the ROM from saves/. This is sensible, but I should add
an option to specifiy which .sav to read from and which .sav to write to. I
should also clean up the memory class before adding more functionality to this.
For now, this functionality was added to mem.cpp, but it might make more sense
to add it to mbc.cpp