significantly improved lcd::step_lcd() performance
Made lcd::pixels store color structs instead of individual bytes.
SDL_UpdateTexture() expects an array of bytes, but if the colors in the struct
are in the correct order then they'll be in the correct order in memory and
the array of colors structs will be identical to the array of bytes. This
improves performance since step_lcd only has to do one assignment per pixel.
Also had to change mem::update_palette() to store the colors in the palette
structs in the correct order. Another somewhat minor performance improvement
came from reordering some loops in case 3 of the switch case statement in
step::lcd() to reduce unnecessary calls to mem::get_palette().