reduced unnecessary calls to mem::read_byte() in mem.cpp
since memory bank switching slightly increased the complexity of
mem::read_byte() it no longer makes sense to use it when the address is
guaranteed to be outside the rom or external ram space. This is really only an
issue when the function(s) accessing memory are called many times. Since the
functions in mbc.cpp which use mem::read_byte() are only called once, the
added overhead from using mem::read_byte() is minimal.