fixed window drawing bug in pkblue
w_offset should be uint16 since it's used as an index in the 32x32 tile map. I
may have inadvertantly introduced another bug that's visible when a program
tries to draw both the background and the window, but that should be an easy fix
once I move away from using lcd::linebuffer and make a lcd::draw_bg() and
lcd::draw_win().