figured out how to manipulate "global" variables from libjit
see https://lists.gnu.org/archive/html/libjit/2016-06/msg00001.html
I mean global variables in the sense that they are normally outside the scope of
the libjit functions. basically I have to make a type for the address of the
variables with jit_type_create_pointer(). Then make a jit_value containing the
address of the variable using new_constant() or similar. Finally use
insn_load_relative() and insn_store_relative() to modify the variable. This is
more complicated than it should be but it should work for now.