correctly implemented opcodes 0x01 (ld bc xx) and 0xc5 (push bc)
the previous commit did not include the 'this' argument to mem::read_word.
when calling a non-static member function using a pointer there is a hidden
argument (this) specifying the instance of the class to call the method on.
the hidden argument must be explicitly included when calling a non-static
member function through libjit's insn_call_native() and its signature must
reflect that as well. also cleaned up the dynarec main loop and fixed some
logical errors. renamed instructions.h to translations.h since only opcodes
that will be translated to libjit functions are included there; the remaining
opcodes are included in jumps.h and cond_branches.h