added a specific jit_context to each instance of cache_block
this will simplify evicting blocks from the cache since deleting individual
functions in a jit_context is non-trivial. there was some email thread on the
libjit archives about exposing _jit_function_destroy() to enable individual
function deletion, but a custom memory allocator is required. this may or may
not be particularly difficult but a thorough understanding of memory allocation
in libjit is required to add this feature