simplified the cache and got dynarec working with jit_test.asm
when the cache is full of valid blocks, the first block is arbitrarily chosen to
be replaced. the cache also now uses an array of cache_block pointers instead of
a vector. this is ok for now, but ideally I should keep track of another array to
keep track of block priorities so that the last used block can found when we need
to replace a block