replaced address with struct combining address and bank number in cache blocks
using the `location` struct will make it much easier to allow the dynarec to
support ROMs larger than 32 kB. Although the bank number in the struct can
refer to either a ROM bank or RAM bank, it can be inferred from the address.
`dynarec_cpu::translate()` correctly sets the start and end locations of each
block, but will not handle blocks that cross bank boundaries yet. I should
implement this by adding a macro similar to JIT_INVALIDATE_BLOCKS to check for
bank switches. These writes should trigger bank switches both just-in-time and
ahead-of-time to include instructions from multiple banks in one block.
`cache::invalidate_blocks()` was also modified to take a `location` as an
argument, but it completely ignores the bank attribute for now.