finished implementing 0xcb opcodes
All opcodes used in demo.gb are now implemented. When running demo.gb with the
dynarec the ball appears in the middle of the screen as expected, but doesn't
move and the bar doesn't appear yet. I should use the interpreter and the methods
in states.cpp to find issues in the dynarec.