added option to set test mode at runtime and fixed 3 bugs in the dynarec
In test mode we make an instance of cpu and dynarec_cpu and two instances of
each of the aux classes. The dynarec_cpu is run first with a modified emulate()
method that takes a number of steps to run. Since blocks can have more than one
instruction, it may execute more steps than we request so we return the actual
number of instructions executed. This number is then passed to cpu's modified
emulate() method which also takes a number of instructions to execute. After
running this number of steps, we print out the last ten interpreted instructions
and compare the final cpu states. Since the interpreter executes instructions
one at a time, we are guaranteed to always stop both cpu's at the same point.

I could clean up the test mode stuff quite a bit, but it's usable enough for
now. I also found and fixed 3 bugs in the dynarec cpu JIT macros with demo.gb